+++ snowddl/resolver/schema.py
@@ -77,12 +77,6 @@ class SchemaResolver(AbstractResolver):
 
         return ResolveResult.DROP
 
-    def destroy(self):
-        # Schemas are normally dropped automatically on DROP DATABASE
-        # But in some cases explicit DROP SCHEMA might be required, e.g. for SingleDB mode
-        if self.engine.settings.destroy_schemas:
-            super().destroy()
-
     def _post_process(self):
         for result in self.resolved_objects.values():
             if result != ResolveResult.NOCHANGE:
@@ -96,6 +90,10 @@ class SchemaResolver(AbstractResolver):
         for schema_full_name in sorted(self.existing_objects):
             # Object exists in blueprints, should not be dropped
             if schema_full_name in self.blueprints:
+                continue
+
+            # Parent object is going to be dropped
+            if self.engine.intention_cache.check_parent_drop_intention(self.object_type, schema_full_name):
                 continue
 
             database_full_name = ".".join(schema_full_name.split(".")[:1])